+++ packages/cli/uni-builder/src/shared/plugins/devtools.ts
@@ -19,11 +19,13 @@ export const pluginDevtool = (options: {
     // priority order
     // 1. output.sourceMap.js, if this value is set, we won't apply this plugin and let rsbuild handles it
     const devtoolJs =
-      typeof options.sourceMap === 'boolean'
-        ? options.sourceMap
-        : options.sourceMap?.js;
+      typeof options.sourceMap === 'boolean' ||
+      options.sourceMap?.js !== undefined;
     if (devtoolJs) {
-      if (!isUseJsSourceMap(options.disableSourceMap)) {
+      if (
+        typeof options.disableSourceMap === 'boolean' ||
+        options.disableSourceMap?.js !== undefined
+      ) {
         logger.warn(
           'Detected that `output.sourceMap` and `output.disableSourceMap` are used together, use the value of `output.sourceMap`',
         );