+++ packages/cli/plugin-proxy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/plugin-proxy
 
+## 2.63.0
+
+### Patch Changes
+
+- @modern-js/utils@2.63.0
+
 ## 2.62.1
 
 ### Patch Changes